Conflicts:
	apps/movie-db/src/app/app.root.ts
	libs/movie/feature-media-items/src/lib/media-items.page.html
	libs/movie/feature-media-items/src/lib/media-items.page.ts
	libs/movie/ui/media-card/media-card.component.ts
	libs/shared/test/msw/src/assets/mockServiceWorker.js
	package.json
	pnpm-lock.yaml

+++ decorate-angular-cli.js
@@ -30,7 +30,7 @@ try {
   output = require('@nx/workspace').output;
 } catch (e) {
   console.warn(
-    'Angular CLI could not be decorated to enable computation caching. Please ensure @nx/workspace is installed.'
+    'Angular CLI could not be decorated to enable computation caching. Please ensure @nx/workspace is installed.',
   );
   process.exit(0);
 }
@@ -49,8 +49,7 @@ function symlinkNgCLItoNxCLI() {
        * Such that it works in all shells and works with npx.
        */
       ['', '.cmd', '.ps1'].forEach((ext) => {
-        if (fs.existsSync(nxPath + ext))
-          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
+        if (fs.existsSync(nxPath + ext)) fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
       });
     } else {
       // If unix-based, symlink
